+++ 03.asynchronous/lib/callback/success_callback.js
