+++ 03.asynchronous/lib/promise/success_promise.js
@@ -5,7 +5,7 @@ import {
   dbRunPromise,
   dbAllPromise,
   dbClosePromise,
-} from "../module/sqlite3_functions.js";
+} from "../../module/sqlite3_functions.js";
 
 function main() {
   const db = new sqlite3.Database(":memory:");
