+++ 03.asynchronous/lib/promise/unsuccess_promise.js
@@ -6,7 +6,7 @@ import {
   dbRunPromise,
   dbAllPromise,
   dbClosePromise,
-} from "../module/sqlite3_functions.js";
+} from "../../module/sqlite3_functions.js";
 
 function main() {
   const db = new sqlite3.Database(":memory:");
